Moved DependencyLoader to util.maven package, and renamed the loader class. Fixed stuff, so it now works.
