+++ orbis-core/src/main/java/com/azortis/orbis/util/maven/DependencyLoader.java
@@ -16,7 +16,7 @@
  *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.azortis.orbis.depend;
+package com.azortis.orbis.util.maven;
 
 import com.azortis.orbis.Orbis;
 import org.slf4j.Logger;
@@ -25,7 +25,6 @@ import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.MalformedURLException;
-import java.net.URISyntaxException;
 import java.net.URL;
 import java.net.URLClassLoader;
 import java.nio.file.Files;
@@ -33,7 +32,7 @@ import java.nio.file.Files;
 /**
  * A class that handles downloading of dependencies for the plugin as well as loading dependencies into the plugin.
  */
-public class DownloadManager {
+public class DependencyLoader {
     private static final AccessLoader INJECTOR;
     private static final Logger logger;
 
@@ -75,14 +74,14 @@ public class DownloadManager {
      * @param dependency The dependency to load.
      */
     public static void load(Dependency dependency) {
-        logger.info(String.format("Loading dependency %s:%s:%s from %s", dependency.group(),
+        logger.debug(String.format("Loading dependency %s:%s:%s from %s", dependency.group(),
                 dependency.artifact(), dependency.version(), dependency.repository().url()));
         String name = dependency.artifact() + "-" + dependency.version();
 
         File saveLocation = new File(getLibFolder(), name + ".jar");
         if (!saveLocation.exists()) {
             try {
-                logger.info(String.format("Dependency %s is not already in the libraries folder. Attempting to download...", name));
+                logger.info(String.format("Dependency %s is not present in the libraries folder. Attempting to download...", name));
                 URL url = getUrl(dependency);
                 try (InputStream stream = url.openStream()) {
                     Files.copy(stream, saveLocation.toPath());
@@ -104,7 +103,7 @@ public class DownloadManager {
             throw new RuntimeException("Unable to load dependency: " + name, e);
         }
 
-        logger.info(String.format("Dependency %s has been loaded", name));
+        logger.debug(String.format("Dependency %s has been loaded", name));
     }
 
 
@@ -115,17 +114,12 @@ public class DownloadManager {
      */
     public static File getLibFolder() {
         File jarFile;
-        try {
-            jarFile = new File(Orbis.getPlatform().getMainClass().getProtectionDomain().getCodeSource().getLocation().toURI());
-            File libFolder = new File(jarFile.getParentFile(), "libs");
-            if (!libFolder.exists()) {
-                libFolder.mkdirs();
-            }
-            return libFolder;
-        } catch (URISyntaxException e) {
-            e.printStackTrace();
+        jarFile = new File(Orbis.getPlatform().directory() + "/libs/");
+        File libFolder = new File(jarFile.getParentFile(), "libs");
+        if (!libFolder.exists()) {
+            libFolder.mkdirs();
         }
-        return null;
+        return libFolder;
     }
 
     /**
